CargoHubV2/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
sqlite3